Restructure code to exclude political capital and to include a more detailed model of the RChain state

+++ simulation/src/main/scala/coop/rchain/caspersimulation/reporting/RevFlow.scala
@@ -2,9 +2,9 @@ package coop.rchain.caspersimulation.reporting
 import coop.rchain.caspersimulation.Validator
 import coop.rchain.caspersimulation.network.Network
 
-object RevFlow extends ValidatorScalarFlow[Double] {
+object RevFlow extends ValidatorScalarFlow[Int] {
 
-  override def observe(input: Network): Map[Validator, Double] = {
+  override def observe(input: Network): Map[Validator, Int] = {
     input.validators.map(v => v -> v.revEarned).toMap
   }
